+++ src/components/Debug.svelte
@@ -4,22 +4,30 @@
   import Label from "./AnswerTable/Label.svelte";
   import { CaptchaType } from "../models/CaptchaType";
   import { trimHTML } from "../utils/Trim";
+  import { Insight } from "../models/Insight";
 
   let data: { [category: string]: CaptchaType } = {};
+  let insight: Insight;
 
   onMount(async () => {
-    const res = await fetch("/api/getall");
-    data = await res.json();
+    const res1 = fetch("/api/getall");
+    const res2 = fetch("/api/insight");
+    insight = await (await res2).json();
+    data = await (await res1).json();
   });
+
+  const I = (category: string) => insight.categories[category];
 </script>
 
 <main>
+  <p class="mt-4">{insight ? `Data at ${insight.created_at}` : "Loading..."}</p>
   {#each Object.entries(data) as [category, captcha]}
     <h1 class="display-6">Category: {category}</h1>
     <h2>{trimHTML(captcha.title)}</h2>
     <table class="table table-bordered">
       <thead>
         <tr class="table-info">
+          <th>Category</th>
           <th>Index</th>
           <th>Image</th>
           <th>Character Name</th>
@@ -29,6 +37,7 @@
       <tbody>
         {#each captcha.questions as question, i}
           <tr class={question.answer ? "table-success" : "table-danger"}>
+            <td><span class="fs-4">{category}</span></td>
             <td><h1>{i + 1}</h1></td>
             <td>
               <div
@@ -42,6 +51,12 @@
         {/each}
       </tbody>
     </table>
+
+    <p>
+      <span>correct_count = {I(category).correct_count},</span>
+      <span>questions_count = {I(category).questions_count},</span>
+      <span>correct_ratio = {I(category).correct_ratio}</span>
+    </p>
   {/each}
 </main>
 
@@ -57,7 +72,8 @@
     background-position: center;
   }
 
-  th {
+  th,
+  p {
     font-size: 1.3em;
   }
 </style>